requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
huggingface_hub==0.15.1
pyodbc==5.0.1
requests==2.31.0
python-dotenv==1.0.1
reportlab==4.0.8
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.4
scipy==1.10.1
certifi>=2023.7.22
groq==0.3.0
